fix a error of del_pkg_rpm function

+++ script/tools/pkg_rpm_update.sh
@@ -235,7 +235,11 @@ function del_pkg_rpm(){
 	do
 		osc ls -b ${obs_proj} ${pkg} standard_aarch64 aarch64 2>/dev/null | grep rpm > aarch_rpmlist.txt
 		osc ls -b ${obs_proj} ${pkg} standard_x86_64 x86_64 2>/dev/null | grep rpm > x86_rpmlist.txt
-		src_rpm=`cat aarch_rpmlist.txt | grep "src.rpm"`
+		if [ -s aarch_rpmlist.txt ];then
+			src_rpm=`cat aarch_rpmlist.txt | grep "src.rpm"`
+		else
+			src_rpm=`cat x86_rpmlist.txt | grep "src.rpm"`
+		fi
 		rpm_name=`echo ${src_rpm%%-[0-9]*}`
 		tmp=`echo ${src_rpm%%.oe1*}`
 		version=`echo ${tmp#*-}`
@@ -245,16 +249,20 @@ function del_pkg_rpm(){
 		big_version=`echo ${version%%-*}`
 		sed -i '/src.rpm/d' aarch_rpmlist.txt
 		sed -i '/src.rpm/d' x86_rpmlist.txt
-		for line in `cat aarch_rpmlist.txt`
-		do
-			name=`echo ${line%%-[0-9]*}`
-			ssh -i ${update_key} -o StrictHostKeyChecking=no root@${update_ip} "cd ${pkg_aarch_path} && ls ${name}-*.rpm 2>/dev/null | grep ${big_version} | xargs rm 2>/dev/null"
-		done
-		for line2 in `cat x86_rpmlist.txt`
-		do
-			name2=`echo ${line2%%-[0-9]*}`
-			ssh -i ${update_key} -o StrictHostKeyChecking=no root@${update_ip} "cd ${pkg_x86_path} && ls ${name2}-*.rpm 2>/dev/null | grep ${big_version} | xargs rm 2>/dev/null"
-		done
+		if [ -s aarch_rpmlist.txt ];then
+			for line in `cat aarch_rpmlist.txt`
+			do
+				name=`echo ${line%%-[0-9]*}`
+				ssh -i ${update_key} -o StrictHostKeyChecking=no root@${update_ip} "cd ${pkg_aarch_path} && ls ${name}-*.rpm 2>/dev/null | grep ${big_version} | xargs rm 2>/dev/null"
+			done
+		fi
+		if [ -s x86_rpmlist.txt ];then
+			for line2 in `cat x86_rpmlist.txt`
+			do
+				name2=`echo ${line2%%-[0-9]*}`
+				ssh -i ${update_key} -o StrictHostKeyChecking=no root@${update_ip} "cd ${pkg_x86_path} && ls ${name2}-*.rpm 2>/dev/null | grep ${big_version} | xargs rm 2>/dev/null"
+			done
+		fi
 		ssh -i ${update_key} -o StrictHostKeyChecking=no root@${update_ip} "cd ${source_path} && ls ${rpm_name}-*.src.rpm 2>/dev/null | grep ${big_version} | xargs rm 2>/dev/null"
 	done
 	rm -f *_rpmlist.txt
